Add files via upload
*Please Read before Executing*
Project 2 contains the entirety of codes used for this project.
algorithm.R contains only the algorithm required to produce predictions. It requires the user to set the path to trainingdata and testdata.
trainingdata.csv is the training data I used for modelling and creating algorithm. Please load this file into algorithm.R